First draft for code arrangement plus first draft contract design
I've created the src directory which will contain all the code for the project.
Inside the backend directory will be all the smart contracts, deploy scripts and tests
Inside the front end will be all the websites code

Note that the backend code for now is not definitive is just the first draft as an intuitive
approach on how to continue developing the code. Interfaces are just a demonstration they are not correct.